+++ src/main/java/uk/co/transactis/Shape.java
@@ -1,12 +1,5 @@
 package uk.co.transactis;
 
-/**
- * Created with IntelliJ IDEA.
- * User: martin
- * Date: 06/04/14
- * Time: 08:44
- * To change this template use File | Settings | File Templates.
- */
 public interface Shape {
-    public float getArea();
+    public double getArea();
 }
